4.21*: Blog List Expansion, step 9

+++ part4/blog/controllers/blogs.js
@@ -32,7 +32,25 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  console.log("blog_id: ", blog.user.toString())
+  console.log("user_id: ", user.id.toString())
+
+  if ( blog.user.toString() === user.id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+  } else {
+    return response.status(401).json({ error: 'not authorized' })
+  }
+  
   response.status(204).end()
 })
